requirements.txt
# Core dependencies
langchain==0.1.16
langchain-openai==0.0.8
langchain-anthropic==0.1.0
langchain-community==0.0.32
langchain-core==0.1.42
openai==1.10.0

# API framework
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database and caching
sqlalchemy==2.0.36
alembic==1.14.0
redis==5.2.0
asyncpg==0.30.0

# Vector store
chromadb==0.5.0
faiss-cpu==1.9.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.0
tenacity==8.2.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
structlog==24.4.0
prometheus-client==0.21.0

# Data processing
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2

# Development tools
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

## Additional dependencies added as part of debbugging from claude code
slowapi
greenlet
